# Refactor best YouTube URL and add `yt_dlp` option