--- a/ultralytics/yolo/data/dataloaders/stream_loaders.py
+++ b/ultralytics/yolo/data/dataloaders/stream_loaders.py
@@ -45,9 +45,17 @@ class LoadStreams:
             st = f'{i + 1}/{n}: {s}... '
             if urlparse(s).hostname in ('www.youtube.com', 'youtube.com', 'youtu.be'):  # if source is YouTube video
                 # YouTube format i.e. 'https://www.youtube.com/watch?v=Zgi9g1ksQHc' or 'https://youtu.be/Zgi9g1ksQHc'
-                check_requirements(('pafy', 'youtube_dl==2020.12.2'))
-                import pafy  # noqa
-                s = pafy.new(s).getbest(preftype='mp4').url  # YouTube URL
+                check_requirements(('yt-dlp'))
+                import yt_dlp
+                with yt_dlp.YoutubeDL({}) as ydl:
+                    # get all information about the youtube video
+                    info = ydl.extract_info(s, download=False)
+                    formats = info['formats']
+                    # the last listed mp4 format is the best one.
+                    for format in reversed(formats):
+                        if (format['ext'] == 'mp4'):
+                            s = format['url']
+                            break
             s = eval(s) if s.isnumeric() else s  # i.e. s = '0' local webcam
             if s == 0 and (is_colab() or is_kaggle()):
                 raise NotImplementedError("'source=0' webcam not supported in Colab and Kaggle notebooks. "