Pafy is outdated, use yt-dln to get better mp4 urls.

Fixes: #1837 

I have read the CLA Document and I sign the CLA
<!--
copilot:all
-->
### <samp>🤖 Generated by Copilot at b8cf5a8</samp>

### Summary
🔄🎥🛠️

<!--
1.  🔄 This emoji means "repeat" or "update" and can be used to indicate that the code has been updated to use a different library or tool for the same functionality.
2.  🎥 This emoji means "movie camera" or "video" and can be used to indicate that the code deals with video sources or formats, such as YouTube videos or mp4 files.
3.  🛠️ This emoji means "hammer and wrench" or "tools" and can be used to indicate that the code fixes a bug or issue, such as the inability to load some videos as streams.
-->
Use yt-dlp instead of pafy and youtube_dl to stream YouTube videos in `ultralytics/yolo/data/dataloaders/stream_loaders.py`. This fixes a bug that prevented some videos from being loaded.

> _`yt-dlp` replaces_
> _old libraries for YouTube_
> _streaming in winter_

### Walkthrough
* Replace pafy and youtube_dl with yt-dlp to extract best mp4 URL from YouTube videos ([link](https://github.com/ultralytics/ultralytics/pull/2901/files?diff=unified&w=0#diff-be9cf5c74e77e0cde8d630f233c18b23147891ede552e737c1ad702c08f7b652L48-R58))