@tekert 2023-05-30
Made some changes to use urls with no audio and put all in a one liner.