@glenn-jocher 2023-05-30
@tekert thanks for the PR! I've updated this a bit but discovered problems with the URL returned by the PR. The current solution produces a correct result like this on our default video:

### main branch
```
YOLOv8n summary (fused): 168 layers, 3151904 parameters, 0 gradients, 8.7 GFLOPs

1/1: https://youtu.be/Zgi9g1ksQHc... Success ✅ (6557 frames of shape 1280x676 at 24.00 FPS)

0: 352x640 (no detections), 46.1ms
0: 352x640 (no detections), 46.0ms
0: 352x640 (no detections), 40.5ms
0: 352x640 (no detections), 42.4ms
0: 352x640 (no detections), 36.8ms
0: 352x640 (no detections), 41.3ms
0: 352x640 (no detections), 36.6ms
0: 352x640 (no detections), 35.6ms
```


### this PR
Pulls very low resolution video which becomes unresponsive after a few frames.
```
YOLOv8n summary (fused): 168 layers, 3151904 parameters, 0 gradients, 8.7 GFLOPs

1/1: https://youtu.be/Zgi9g1ksQHc... Success ✅ (3278 frames of shape 256x136 at 12.00 FPS)

0: 352x640 4 persons, 54.9ms
0: 352x640 3 persons, 53.5ms
0: 352x640 1 person, 51.9ms
0: 352x640 1 person, 51.3ms
WARNING ⚠️ Video stream unresponsive, please check your IP camera connection.

```