@glenn-jocher 2023-05-30
@tekert I'll merge this in any case with pafy still as the default, but if you could work on the yt_dln side so it matches the pafy URL, then we can switch the default use_pafy=False.