@glenn-jocher 2023-05-30
@tekert bonus points if you can figure out how to stream the next highest resolution above the inference `imgsz`. i.e. for imgsz=640 would be awesome to stream at 640 directly instead of 1280.