@tekert 2023-05-30
Hi glenn, you removed the `reversed` method when iterating, the last format in the dictionary is the best one.
Thats why you get the worst one when you refactored.
But
For the `imgsz`, yes, we have this to play with:
For example the `formats` dict for the best mp4 no audio version in https://youtu.be/Zgi9g1ksQHc

```json
{
    "asr": null,
    "filesize": 83639148,
    "format_id": "137",
    "format_note": "1080p",
    "source_preference": -1,
    "fps": 24,
    "audio_channels": null,
    "height": 1012,
    "quality": 9.0,
    "has_drm": false,
    "tbr": 2449.098,
    "url": "<THE URL>",
    "width": 1920,
    "language": null,
    "language_preference": -1,
    "preference": null,
    "ext": "mp4",
    "vcodec": "avc1.640028",
    "acodec": "none",
    "dynamic_range": "SDR",
    "vbr": 2449.098,
    "protocol": "http_dash_segments",
    "container": "mp4_dash",
    "resolution": "1920x1012",
    "aspect_ratio": 1.9,
    "http_headers": {
        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/93.0.4577.82 Safari/537.36",
        "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8",
        "Accept-Language": "en-us,en;q=0.5",
        "Sec-Fetch-Mode": "navigate"
    },
    "video_ext": "mp4",
    "audio_ext": "none",
    "format": "137 - 1920x1012 (1080p)"
}
```
For imgz we can check videos width <= imgz if not found fallback to the best mp4 one.

I'll updated the branch with a posible solution for this.