@glenn-jocher 2023-05-31
@tekert hi! Thank you for your message. Yes, the reversed method was removed when iterating, but the order of the dictionary doesn't change the fact that the first format is not always the best quality. We use the pafy library to get the URLs for the best mp4 videos without audio because it is known to provide the best quality. However, we are open to using other libraries if they provide better quality URLs.

Regarding the `imgsz`, we use the `formats` dict to get the best mp4 no audio version based on the resolution criteria you provided. Your proposed solution to check video width <= imgz is interesting and we look forward to seeing it. Please submit a pull request so we can review the code changes.

Thank you for your contributions to the YOLOv8 project!