@tekert 2023-05-31
No problem glad to help. Continued on #2914

True, the are progressive (video+audio) videos, and adaptative (dash videos), the dash videos are demuxed meaning video and audio streams are separated, pafy when using "best" is always using the best progressive video wich there are less formats available. I just used the last format listed wich uses lang,quality,res,fps,.... and so on as order (wich can be changed) is a better library than most.

Random toughts:
While working on this i just got the weird idea to form a dataset with inferred yolo segmentations from youtube, automattically convert them to annotations, review and correct them manually (humans can detect this very fast more than half the work is done)  and train again, i think i can improve some things. Would be nice to infer the context of the images too. And change the model to use previous same context infered images as a input of some form for new (same context) images.