@glenn-jocher 2023-06-01
Thank you @tekert for providing the explanation about the different types of videos and the behavior of pafy when using "best". Your proposed solution of using the last format listed is interesting and we appreciate your efforts in exploring alternatives to pafy. 

Regarding your random thought, creating a dataset with inferred YOLO segmentations from YouTube videos sounds like an interesting idea. It would definitely be helpful to have a dataset that includes annotations to train the model. However, we recommend reviewing the YouTube terms of service to ensure that it is permissible to use their videos for this purpose. Thank you for your contributions to the YOLOv8 project!